feat(codemirror): add working JSON validator
