+++ package.json
@@ -13,6 +13,7 @@
     "html-webpack-plugin": "^4.5.0",
     "husky": "^4.3.0",
     "lint-staged": "^10.4.0",
+    "process": "^0.11.10",
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
@@ -21,7 +22,8 @@
     "sass": "^1.26.11",
     "sass-loader": "^10.0.2",
     "semantic-commitlint": "^1.4.0",
-    "source-map-loader": "^1.1.0"
+    "source-map-loader": "^1.1.0",
+    "stream-browserify": "^3.0.0"
   },
   "scripts": {
     "start": "npx webpack-dev-server",
