+++ src/components/customizers/JsonCustomizer.js
@@ -1,9 +1,11 @@
-import React, { useRef, useEffect, useState } from 'react';
-import Codemirror from 'codemirror';
-import data from '../../example-json/john_smith.json';
-import { editorOptions } from './EditorOptions';
-import PropTypes from 'prop-types';
+import CodeMirror from 'codemirror';
 import 'jsonlint';
+import PropTypes from 'prop-types';
+import React, { useRef, useEffect, useState } from 'react';
+
+import data from '../../example-json/john_smith.json';
+import './lint';
+import { editorOptions } from './options';
 
 export const JsonCustomizer = ({ setData }) => {
   const [editor, setEditor] = useState(null);
@@ -15,13 +17,13 @@ export const JsonCustomizer = ({ setData }) => {
   };
 
   useEffect(() => {
-    setEditor(Codemirror.fromTextArea(editorRef.current, editorOptions));
+    setEditor(CodeMirror.fromTextArea(editorRef.current, editorOptions));
     setData(data);
   }, [editorRef]);
 
   return (
     <div>
-      <button onClick={handleClick}>Click me</button>
+      <button onClick={handleClick}>Update PDF</button>
       <form>
         <textarea
           onChange={onHandleChange}
